requirements.txt
Click==7.0
coloredlogs==10.0
Flask==1.0.2
humanfriendly==4.17
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
pywebview==2.2.1
Werkzeug==0.14.1
